backend/requirements.txt
fastapi
uvicorn
python-dotenv
aiosmtplib
python-multipart
motor
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
httpx
razorpay